modified:   src/components/RightSideBar.jsx 	renamed:    src/components/SearchBar.jsx -> src/components/RightSideBar/SearchBar.jsx 	renamed:    src/components/SubscriptionBox.jsx -> src/components/RightSideBar/SubscriptionBox.jsx 	renamed:    src/components/TrendingItem.jsx -> src/components/RightSideBar/TrendingItem.jsx 	renamed:    src/components/WhatsHappening.jsx -> src/components/RightSideBar/WhatsHappening.jsx
