+++ src/components/RightSideBar/SearchBar.jsx
@@ -9,7 +9,7 @@ const SearchBar = () => {
         className="w-full p-3 pl-10 rounded-full bg-slate-950 text-white placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-blue-500 border-2 border-gray-700"
       />
       <svg
-        className="absolute left-3 top-3 w-5 h-5 text-gray-400"
+        className="absolute mt-1 left-3 top-3 w-5 h-5 text-gray-400"
         xmlns="http://www.w3.org/2000/svg"
         fill="none"
         viewBox="0 0 24 24"
